pyproject.toml
[project]
name = "ml-template"
version = "0.1.0"
description = "A machine learning project template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "torchvision",
    "lightning",
    "tensorboard",
    "pillow",
    "matplotlib",
    "numpy",
    "opt-einsum",
    "hydra-core",
    "pandas",
    "torch-tb-profiler",
    "wandb",
    "scipy",
]

[dependency-groups]
dev = [
    "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
